feat: Add GTFS agency and feed_info files
This commit adds the GTFS agency.txt and feed_info.txt files to the project. These files contain information about the transit agency and the feed information, such as the publisher name, URL, language, and start/end dates. The addition of these files will enhance the functionality of the project.

Note: This commit message follows the conventional format of "feat: <description>".